backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
httpx>=0.27.0
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.1,<4.2.0
huggingface_hub>=0.20.0
Pillow>=10.0.0
